Moved scheduler out of main
